ingest_to_cloud_storage/requirements.txt
# Function dependencies, for example:
# package>=version
pandas==1.3.5
google-cloud-storage==2.5.0
functions-framework==3.*